test: address Copilot review on the teacher swap (#117)
- teacher.py: clarify openr1_distill_7b docstring — HF config.json has no
  tokenizer_vocab_size, so from_pretrained(config=None) emits over the padded
  model vocab; pass this config (or set tokenizer_vocab_size) for real loads.
- move the portable TeacherConfig shape/validate tests to a new no-MLX module
  (tests/test_teacher_config.py) so they run on non-Apple/CI hosts where the
  MLX-guarded test_teacher.py is skipped, actually protecting the fixtures; add
  a from_hf_dict test locking in the documented tokenizer_vocab_size gotcha.
